# api_docs.yaml
openapi: 3.0.0
info:
  title: Micro Cold Spray API
  description: API documentation for the Micro Cold Spray system
  version: 1.0.0
  contact:
    name: Support Team
servers:
- url: http://localhost:8000
  description: UI Service
- url: http://localhost:8001
  description: Config Service
- url: http://localhost:8002
  description: State Service
- url: http://localhost:8003
  description: Communication Service
- url: http://localhost:8004
  description: Process Service
- url: http://localhost:8005
  description: Data Collection Service
- url: http://localhost:8006
  description: Validation Service
components:
  schemas:
...
    ConfigRequest:
      type: object
      required:
      - name
      - data
      - format
      properties:
        name:
          type: string
          description: Configuration name
        data:
          type: object
          description: Configuration data
        format:
          type: string
          description: Data format (yaml/json)
          enum: [ yaml, json ]
    ConfigResponse:
      type: object
      required:
      - name
      - data
      - format
      properties:
        name:
          type: string
          description: Configuration name
        data:
          type: object
          description: Configuration data
        format:
          type: string
          description: Data format (yaml/json)
    SchemaRequest:
      type: object
      required:
      - name
      - schema_definition
      properties:
        name:
          type: string
          description: Schema name
        schema_definition:
          type: object
          description: JSON Schema definition
    SchemaResponse:
      type: object
      required:
      - name
      - schema_definition
      properties:
        name:
          type: string
          description: Schema name
        schema_definition:
          type: object
          description: JSON Schema definition
    MessageResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          description: Response message
...
paths:
...
  /config/{name}:
    get:
      summary: Get Configuration
      description: Get configuration by name
      operationId: getConfig
      tags: [ Config ]
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Configuration details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '500':
          description: Internal server error
    put:
      summary: Update Configuration
      description: Update configuration
      operationId: updateConfig
      tags: [ Config ]
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigRequest'
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal server error
  /config/validate/{name}:
    post:
      summary: Validate Configuration
      description: Validate configuration against schema
      operationId: validateConfig
      tags: [ Config ]
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigRequest'
      responses:
        '200':
          description: Configuration validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation failed
  /config/schema/{name}:
    get:
      summary: Get Schema
      description: Get schema by name
      operationId: getSchema
      tags: [ Config ]
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Schema details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaResponse'
        '404':
          description: Schema not found
        '500':
          description: Internal server error
    put:
      summary: Update Schema
      description: Update schema
      operationId: updateSchema
      tags: [ Config ]
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaRequest'
      responses:
        '200':
          description: Schema updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal server error
...